service

+++ src/main/java/com/usman/forum/service/Implementation/AnswerImp.java
@@ -3,9 +3,11 @@ package com.usman.forum.service.Implementation;
 import com.usman.forum.exception.BusinessException;
 import com.usman.forum.model.Answers;
 
+import com.usman.forum.model.Likes;
 import com.usman.forum.model.Questions;
 import com.usman.forum.model.User;
 import com.usman.forum.repository.AnswerRepository;
+import com.usman.forum.repository.LikesReposiory;
 import com.usman.forum.repository.QuestionRepository;
 import com.usman.forum.repository.UserRepository;
 import com.usman.forum.service.AnswersService;
@@ -33,14 +35,18 @@ public class AnswerImp  implements AnswersService {
 
     private final AnswerRepository answerRepository;
     private final QuestionRepository questionRepository;
-    private  final UserRepository userRepository;
+
+    private  final UserImp userImp;
+    private  final  QuestionImpl questionImpl;
+    private  final LikesReposiory likesReposiory;
 
     @Override
     public void saveAnswer(@Valid Answers answers , Long uid, Long questionid) {
 
 
-        User user=userRepository.findById(uid).get();
-        Questions question =questionRepository.findById( questionid).get();
+        User user=userImp.findUser(uid);
+
+        Questions question =questionImpl.findAQuestion( questionid);
 
         Optional<Answers> alreadyAnswered= answerRepository.findAnswerByQuestionAndUser(questionid,uid);
         if( alreadyAnswered.isPresent()){
@@ -50,6 +56,7 @@ public class AnswerImp  implements AnswersService {
 
         question.setAnswered(true);
          answers.setUser(user);
+
         answers.setQuestion(question);
 
         answerRepository.save(answers);
@@ -63,7 +70,8 @@ public class AnswerImp  implements AnswersService {
 
     @Override
     public Answers findAnswer(Long id) {
-        return findAnswers(id);
+      return answerRepository.findById(id).
+                orElseThrow(()-> new BusinessException(HttpStatus.NOT_FOUND, "There is  not such Id in our System: "+id));
     }
 
     @Override
@@ -78,26 +86,29 @@ public class AnswerImp  implements AnswersService {
 
     @Override
     public void deleteAnswer(Long userID,Long id, Long questionId) {
-        findUserByID(userID);
+        User user=userImp.findUser(userID);
         Answers answer=findAnswer(id);
+       if(user.equals(answer.getUser())){
         Long isAnswered=answerRepository.numberOfAnsweredQuestion(questionId);
-
         if(isAnswered==1){
-            Questions question=questionRepository.findById(questionId).get();
+            Questions question=questionImpl.findAQuestion(questionId);
             answerRepository.delete(answer);
             question.setAnswered(false);
             questionRepository.save(question);
             return;
         }
         else
-         answerRepository.delete(answer);
+         answerRepository.delete(answer);}
+       else {
+           throw  new BusinessException(HttpStatus.FORBIDDEN, "You have no permission to delete this");
+       }
 
     }
 
     @Override
     public void bestAnswer(Long userID,Long id, Long questionId) {
-        User user=findUserByID(userID);
-        Answers answer=findAnswers(id);
+        User user=userImp.findUser(userID);
+        Answers answer=findAnswer(id);
         List<Answers> answersByQuestionID =answerRepository.findAllAnswersByQuestionID(questionId)
                 .stream()
                 .filter(s-> s.isBestAnswer()).collect(Collectors.toList());
@@ -110,19 +121,37 @@ public class AnswerImp  implements AnswersService {
         answerRepository.save(answer);
     }
 
+    @Override
+    public String likeAnswer(Long userI, Long answerId) {
+        Answers answers= findAnswer(answerId);
+        userImp.findUser(userI);
+        Optional<Likes> likes=likesReposiory.findLikeByAnmswerAndUser(answerId,userI);
 
+        if( likes.isPresent()){
+            likes.get().setUser(userI);
+            likes.get().setAnswer(answerId);
 
+            int increaseLike = answers.getLikeCount() - 1;
+            answers.setLikeCount(increaseLike);
+            likesReposiory.delete(likes.get());
+            answerRepository.save(answers);
+            return "unliked";
+        }
 
-    private final Answers findAnswers(Long id)
-    {
-        Answers answers= answerRepository.findById(id).
-                orElseThrow(()-> new BusinessException(HttpStatus.NOT_FOUND, "There is  not such Id in our System: "+id));
-        return answers;
+            Likes newLike = new Likes();
+            newLike.setUser(userI);
+            newLike.setAnswer(answerId);
+
+            int increaseLike = answers.getLikeCount() + 1;
+            answers.setLikeCount(increaseLike);
+            likesReposiory.save(newLike);
+            answerRepository.save(answers);
+
+        return  "liked";
+
     }
 
-    private User findUserByID(Long id){
-        User user= userRepository.findById(id).
-                orElseThrow(()-> new BusinessException(HttpStatus.NOT_FOUND, "There is  not such Id in our System: "+id));
-        return user;
-    }
+
+
+
 }
